pyproject.toml
[project]
name = "dakaraserver"
authors = [
        {name = "Neraste", email = "neraste.herr10@gmail.com"},
]
description = "Server for the Dakara Project"
readme = "README.md"
license = {file = "LICENSE"}
dynamic = ["version"]
requires-python = ">= 3.9"
classifiers = [
        "License :: OSI Approved :: MIT License",
        "Programming Language :: Python",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Operating System :: OS Independent",
        "Environment :: Console",
        "Intended Audience :: Developers",
]
dependencies = [
    "APScheduler>=3.9.1,<3.10.0",
    "channels>=3.0.5,<3.1.0",
    "coreapi>=2.3.3,<2.4.0",
    "dj-database-url>=1.0.0,<1.1.0",
    "django-cache-lock>=0.2.5,<0.3.0",
    "django-filter>=22.1,<22.2",
    "django-ordered-model>=3.6,<3.7",
    "django-rest-registration>=0.7.2,<0.8.0",
    "Django>=3.2.16,<3.3.0",
    "djangorestframework>=3.13.0,<3.14.0",
    "Markdown>=3.4.1,<3.5.0",
    "packaging>=21.3,<22.0",
    "Pygments>=2.13.0,<2.14.0",
    "python-decouple>=3.6,<3.7",
]

[project.optional-dependencies]
dev = [
    "black>=22.10.0,<22.11.0",
    "codecov>=2.1.12,<2.2.0",
    "freezegun>=1.2.1,<1.3.0",
    "flake8>=7.0.0,<7.1.0",
    "isort>=5.10.1,<5.11.0",
    "pre-commit>=2.20.0,<2.21.0",
    "pytest-asyncio>=0.20.2,<0.21.0",
    "pytest-cov>=4.0.0,<4.1.0",
    "pytest-django>=4.5.2,<4.6.0",
    "pytest-mock>=3.10.0,<3.11.0",
    "pytest>=7.2.0,<7.3.0",
    "ruff>=0.3.0,<0.4.0",
]
# note: update .pre-commit-config.yaml as well

[project.urls]
Homepage = "https://github.com/DakaraProject/dakara-server"
Repository = "https://github.com/DakaraProject/dakara-server.git"
Issues = "https://github.com/DakaraProject/dakara-server/issues"
Changelog = "https://github.com/DakaraProject/dakara-server/blob/master/CHANGELOG.md"

[project.scripts]
"manage.py" = "dakara_server:django_manage"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "dakara_server.version.__version__" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.coverage.run]
omit = ["*/tests/*", "*/conftest.py"]

[tool.pytest]
addopts = "--cov=dakara_server"
asyncio_mode = "strict"

[tool.isort]
profile = "black"
known_first_party = [
    "internal",
    "users",
    "library",
    "playlist",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "B"]
ignore = []
